requirements.txt
datedays==2.2.1
Flask==1.1.4
Flask_Session==0.4.0
Flask_SQLAlchemy==2.5.1
pika==1.3.1
pyserial==3.5
SQLAlchemy==1.4.22
Werkzeug==1.0.1
